refactor: move prompt templates out of copilotService.ts to copilotPrompts.ts
